// frontend/src/app/core/models/user.model.ts
export interface User {
  id: string;
  name: string;
  email: string;
  mana: number;
  level: number;
  createdAt?: Date;
  updatedAt?: Date;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  name: string;
  email: string;
  password: string;
  confirmPassword: string;
}